Adds support for .wav files
This version adds support for uploading .wav files. Since .wav doesn't support embedded cover art in the metadata, it first searches for a user provided cover art with the same file name and if one does not exist, it searches for the cover art using the iTunes API and uses that instead.